termk2: split to files
