+++ tkcore/term/main.cpp
@@ -5,7 +5,7 @@
 #include "utils/Interactive.h"
 #include "utils/term_utils.h"
 #include "termk1.h"
-#include "termk2.h"
+#include "k2/termk2.h"
 #include "key2.h"
 #include "key1.h"
 #include "key_find.h"
@@ -98,7 +98,7 @@ int main(int argc, char **argv) {
                 thekey_v1_term::login(filePath);
                 return;
             case 2:
-                thekey_term_v2::login(filePath);
+                thekey_v2::login(filePath);
                 return;
             default:
                 cerr << "storage version " << storageInfo->storageVersion << " not supported " << filePath
